anchor_spl
Module associated_token

Re-exports

pub use spl_associated_token_account;

Structs

AssociatedToken

Create

CreateBumps

Constants

ID

The const program ID.

Functions

create

create_idempotent

get_associated_token_address

Derives the associated token account address for the given wallet address and token mint

get_associated_token_address_with_program_id

Derives the associated token account address for the given wallet address, token mint and token program id